Decimal => Float

+++ app/models/transaction.rb
@@ -1,5 +1,5 @@
 class Transaction
   include Mongoid::Document
   field :date, :type => Date
-  field :amount, :type => Decimal
+  field :amount, :type => Float
 end
